refactor(suite): Rename suite-common/analytics to suite-common/analytics-redux
