+++ packages/suite/tsconfig.json
@@ -13,7 +13,7 @@
     "include": [".", "e2e", "**/*.json"],
     "references": [
         {
-            "path": "../../suite-common/analytics"
+            "path": "../../suite-common/analytics-redux"
         },
         {
             "path": "../../suite-common/analytics-types"
