+++ suite-common/analytics-redux/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-common/analytics",
+    "name": "@suite-common/analytics-redux",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
