+++ suite-common/analytics-redux/src/redux/analyticsReducer.ts
