+++ suite-common/message-system/.depcheckrc.json
@@ -2,6 +2,6 @@
     "ignore-patterns": ["libDev", "lib"],
     "ignores": [
         "// Todo: This need so be solved. See: https://github.com/trezor/trezor-suite/issues/21552",
-        "@suite-common/analytics"
+        "@suite-common/analytics-redux"
     ]
 }
