+++ suite-common/message-system/src/useExperiment.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 
-import { selectAnalyticsInstanceId } from '@suite-common/analytics';
+import { selectAnalyticsInstanceId } from '@suite-common/analytics-redux';
 
 import { getActiveExperimentGroup, getExperimentGroupByInclusion } from './experimentUtils';
 import {
