+++ suite-common/test-utils/src/mocks.ts
@@ -316,9 +316,9 @@ const getWalletTransaction = (t?: Partial<WalletAccountTransaction>): WalletAcco
     ...t,
 });
 
-// Mocked @suite-common/analytics package used in various tests
+// Mocked @suite-common/analytics-redux package used in various tests
 const getAnalytics = () => {
-    const originalModule = jest.requireActual('@suite-common/analytics');
+    const originalModule = jest.requireActual('@suite-common/analytics-redux');
 
     return {
         __esModule: true, // this property makes it work
