+++ suite-native/analytics/tsconfig.json
@@ -3,7 +3,7 @@
     "compilerOptions": { "outDir": "libDev" },
     "references": [
         {
-            "path": "../../suite-common/analytics"
+            "path": "../../suite-common/analytics-redux"
         },
         {
             "path": "../../suite-common/analytics-types"
