+++ suite-native/module-dev-utils/package.json
@@ -14,7 +14,7 @@
         "@react-navigation/native": "7.1.26",
         "@react-navigation/native-stack": "7.9.0",
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics": "workspace:*",
+        "@suite-common/analytics-redux": "workspace:*",
         "@suite-common/firmware": "workspace:*",
         "@suite-common/message-system": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
