+++ suite-native/module-settings/package.json
@@ -15,7 +15,7 @@
         "@react-navigation/native": "7.1.26",
         "@react-navigation/native-stack": "7.9.0",
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics": "workspace:*",
+        "@suite-common/analytics-redux": "workspace:*",
         "@suite-common/mev": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-constants": "workspace:*",
